Update C++ mangling to reflect existing practice
The AAPCS64 said that:

   void f(int8x8_t)

should be mangled as:

   _Z1fu10__Int8x8_t

(with a 'u', using the namespace for vendor extensions).  But both
GCC and clang mangle it as:

   _Z1f10__Int8x8_t

(without the 'u').  What clang and GCC do is now the de factor ABI
for many platforms, so this patch adds it as an allowable variation.

(We shouldn't just change the base ABI, since that would be unfair on
anyone who has implemented the original specification correctly for
their platform.)