Extend the AAPCS64 to support SVE types and registers
This patch imports the SVE parts of the AAPCS64 from:

   https://developer.arm.com/docs/100986/0000

Most of the text is copied verbatim.  However, I took the opportunity
to replace the term "Scalable Vector Aggregate" with "Pure Scalable Type",
which includes individual predicates and vectors as well as aggregates.
This avoids having to refer to "Scalable Vector Types, Scalable Predicate
Types and Scalable Vector Aggregates" in various places.

I also tried to clarify exactly which composite types are pure
scalable types.