Specify ABI handling for 8.7-A's new FPCR bits.
There are three new bits: AH, FIZ and NEP, in bits 0,1,2 respectively.
This change adds AH and FIZ to the list of bits that can be
dynamically modified on purpose (like rounding modes), and puts NEP
into a new category of 'must be 0'.

Rationale: NEP changes the _shape_ of the affected instructions, in the
sense that it makes them write different amounts of output, and depend
on different amounts of their input. So it's very unlikely that code
generation will be able to be NEP-agnostic: code written to assume
NEP=0 won't do anything useful if run with NEP=1 and vice versa.

But the other two flags change minor details of return values of
existing operations, and might be useful not only to code binary-
translated from an architecture with those details of behaviour, but
also to code recompiled from source, if the source was written in the
expectation of running on that architecture.