[PAuthABI] Remove requirement for position independent code.
The argument for requiring position independence is not considered strong
enough to rule out non-position independent code. Avoiding copy
relocations can be achieved in a different way.

Replace with a requirement to avoid copy relocations. These are problematic
as the static linker creates the storage for the copy in zero-initialised
data. This means the compiler/assembler can't communicate the signing
schema in the contents of the place of the relocation. This is analagous
to communicating the signing schema for GOT generating relocations, except
the set of static relocations that may need to communicate the signing
schema is harder to identify.

Two possible ways to avoid copy relocations are:
- Use GOT generating relocations for accessing imported data.
- Indirect via a RELRO pointer to the imported data.

At the moment I don't think that there is a need for avoiding the
address of an imported function being the address of the PLT entry as
the static linker is not involved in creating the storage for the
pointer.