[PAuthABI] Remove appendix containing the explicit GOT signing schema
The consensus is to use an unsigned RELRO GOT in the initial implementations
with a possibility of generating code to avoid needing a GOT if signing is
needed. With that in mind the idea to encode a signing schema in the addend
is unlikely to be adopted so it can be removed.