[PAuthABI] Allow GNU program properties as a PAuth ABI marking schema
A new program property GNU_PROPERTY_AARCH64_FEATURE_PAUTH is defined in
section .note.gnu.property that can be used as an alternative method for
signing PAuthABI-compatible ELF files.At least in the Bionic linker,
this is easier to implement: the property (and the existing
.note.gnu.property section) can be trivially extracted from a DSO, and
is already happening for other properties (mainly
GNU_PROPERTY_AARCH64_FEATURE_1_AND to check for BTI and PAC-ret). In
contrast, a new section would require traversing all the sections in the
ELF and checking for the section name, which is more complicated to
implement and possibly slower.

Unfortunately, .note.gnu.property is not guaranteed to be understood by
all ELF-based platforms, so we still keep the new section
.note.AARCH64-PAUTH-ABI-tag as the default marking schema.