[Morello] Rework the varargs for Morello.
For AAPCS64 (hybrid), if an argument is a capability or contains
capabilities, pass it indirectly. This avoids having an extra register
spill area. This also allows a capability-aware caller to pass
capabilities to a capability-unaware callee that would only construct a
va_list and pass the result to some other capability-aware code.

For AAPCS64-cap (pure capability ABI), this is reworked to add
extra protection. All variadic arguments are passed in memory
in 16-byte slots (if the type pased is larger than 16 or has
larger aligment requirements, pass it indirectly). The capability
to the variadic arguments is passed in C9, and this capability
will be NULL if there are no variadic arguments.

This protects from the following issues:
 - Caller reads more arguments than were passed
 - Some type mismatches

Note: this causes mismatches between AAPCS64 and AAPCS64-cap.