fix sysvabi64 formatting: fix SVG render issue and add page breaks
- the original sysvabi64-elf-layout.svg file didn't render properly when
rendered in a PDF. The PDF renderer used an alternate text in the SVG which
clips the text, and the size of the SVG was bigger than the size of a page,
which broke the rendering process. The solution was to annotate the .rst file to
resize the SVG file, and to redraw the SVG in Inkscape, which aligns it with the
other SVGs in this repo.

- added page breaks between chapters to align with other documents in the repo