Soft-deprecate GNU_PROPERTY_AARCH64_FEATURE_1_PAC
Users should not rely on this property to enquire about the status
of the protection of an ELF file. If it is present, then yes, the
file has been protected. If it is not present, it can still have
been protected to some extent.

The comment on Custom PLT protection has also been removed. This
protection is seen as a separate feature to PAC-ret, since you could
ask the linker to generate protected PLTs regardless of the presence
of PAC-ret.