Include the ABS64 and ABS32 relocations in Dynamic relocations
These relocations are special since they are both dynamic and static
relocations.  At the start of the "Relocation" section (currently 5.7)
we state that "A well-formed executable file or dynamic shared object
has no static relocations after static linking".  At the moment there is
nothing in this document which mentions that the above relocations are
both dynamic and static, and hence it could be confusing to see such
relocations in an executable if just going based on the ABI
documentation.

This commit adds both relocations to the "Dynamic Relocations" table in
order to make the document self-consistent.  We also include a note
explaining that these relocations are special.

--- Explanation of some decisions:
In the dynamic relocations table I do not mention the ELF64 code for
ABS32 even though there is such a code.  This is because for ELF64 the
ABS32 relocation is not dynamic.

I include a comment about how these relocations are dynamic and static
for extra clarity, even though that is supposed to be implicit in the
fact that they occur in both tables.