[CONTRIBUTING] Add section on extension documents
Some changes to the ABI will require experimentation before the content
stabilises. Adding information to the main ABI is premature in this case.
However it is useful to document this information as an extension to the
ABI to enable further collaboration.

This section outlines the process for submitting an extension, what
criteria Arm use for accepting a contribution and how an extension can
move to the main ABI.