Mention size and alignment of "place" for some relocations
Most relocations act on immediates of instructions and specify the
bits of the calculated value that are put into that instruction.

The "Static Data relocations" table is for data relocations which do
not act on instructions.  These seem to have not explicitly mentioned
the number of bits written by the linker.  They also do not mention
the alignment of the "place" that is written (something which is
written in the aaelf32 document).

Here we mention both the above in the "Comment" field of the
relocation tables.  This differs from where the aaelf32 document
describes the size and alignment of static data simply because I felt
this way would be easier to find the information and would follow the
existing style in the document better.
Am happy to change that if requested.

It appears the GOTREL relocations are not implemented in any linker
implementation so these could not be double-checked against existing
implementations.  Other than that existing linker implementations were
checked to ensure that their behaviour follows the new description.