Relax BTI PLT requirement in aaelf64
The ABI states that all PLT entries must start with BTI when
the DT_AARCH64_BTI_PLT dynamic tag is present. This is too
strict as the static linker can prove that some PLT entries
are never indirectly called so do not need to start with a
BTI. Both LLD and GNU ld do this.

Relax the wording to bring the specification in line with
implementations.

https://github.com/ARM-software/abi-aa/issues/197