[AAELF64][PAUTHABI64] Assign PAuthABI relocation codes.
Now that there will be a concrete implementation of PAuthABI in upstream
LLVM we need to move the PAuthABI relocation codes out of the space
reserved for private experiments to their final codes. These are in
the range of codes already reserved for the PAuthABI in AAELF64.

The description of the operations will remain in the PAuthABI64 as
this remains an extension.