[AAELF64][SYSVABI64] Move dynamic linking contents to sysvabi64
The Elf for the Arm 64-bit Architecture (aaelf64) describes the
processor specific parts of the ELF specification common to all
ELF environments from a statically linked freestanding environment
with no OS to a hosted environment that supports dynamic linking.

The dynamic linking parts that are specific to SystemV style
operating environments such as Linux, and in particular the
GNU properties which are based on a Linux ABI extension are
better documented in the SystemV ABI for Arm 64-bit
Architecture.