[DesignDocuments] Rationale for GNU Properties in sysvabi
Add a design rationale for use of GNU properties as well as
guidelines for how these should be used for properties in
the AArch64 processor space.

Pull request https://github.com/ARM-software/abi-aa/pull/228 moves
the GNU properties and other dynamic section properties specific
to SystemV ABI to the SystemV ABI document.

Arm has typically left metadata in exectuables and shared-libraries
to the platform. Only defining metadata for relocatable objects.
With platforms such as Linux the most frequently run software on
AArch64, Arm needs to document the metadata that it is using for
SystemV platforms.

We have chosen to use GNU properties and to document these in
the sysvabi64.rst document.