[aaelf64] Fix equation for R_AARCH64_GOTPCREL32
The current wording for this reloc is `G(GDAT(S+A))-P` which means the
addend is applied to the symbol we're taking the GOT entry for. What we
actually want (and what's implemented in lld) is something similar to
`R_X86_64_GOTPCREL` where the addend is applied to the location of the
reloc rather than the symbol.