[CLIBABI32] Restore missing reference to EDOM.
I just noticed a paragraph that was grammatically confusing because it
contained a comma with no word before it, and an equality with no
left-hand side.

It looks to me as if the <errno.h> macro name EDOM was accidentally
deleted in both cases. (Rationale: EDOM is an error code closely
related to ERANGE, and matches the claim in the paragraph that its
consensus value is 33 – that's its value on Linux, for example.)

Searching for ERANGE in the rest of the document turned up another two
places where there was a missing word that looked as if it ought to be
"EDOM".

I've no idea how this happened! But it seems like an easy fix.