[aaelf64] Fix PAuthABI relocation details.
There is a typo in the spelling of R_<CLS_ABS64, it is missing a
close angle brackets, and should be R_<CLS>_ABS64.

Similarly there should only be one entry in the dynamic relocation
table, with R_<CLS>_ABS64 retaining the static relocation code but
using the dynamic relocation semantics.

Added a clarifying line to state that the R_<CLS>_ABS64 is both a
static and dynamic relocation.