[pauthabi64] Add note for RAARCH64_AUTH_GOT_ADD_LO12_NC
There is no equivalent for this relocation in the standard ABI
it is used by runtime code to calculate the address of
a GOT slot so it can be used as one of the inputs to an
authenticate instruction. Add a note that this matches up
with the :got_auth_lo12: operator for future reference.

Part of https://github.com/ARM-software/abi-aa/issues/253