[ATOMICSABI64]: Alpha Draft of Atomics ABI
This is the Alpha draft of the ABI for the
"C/C++ Atomics Application Binary Interface Standard for the Arm® 64-bit Architecture"

This document describes the C/C++ Atomics Application Binary Interface for the Arm 64-bit architecture.

This document concerns the valid mappings from C/C++ Atomic Operations to sequences of A64 instructions.
This document does not support Armv7.

For matters concerning the memory model, please consult §B2 of the Arm Architecture Reference Manual.

We focus only on a subset of the C11 atomic operations and their mapping to A64 instructions at the time of writing.

More atomics will be added.

Co-Authored with Wilco Dijkstra (@Wilco1).