Removed OPTIONS argument from the subroutines.
As @rsandifo-arm explained, the same routines cannot be used for
general use of saving/restoring of state enabled by PSTATE.ZA
because of the different expectations on the ZA interface
when partially saving/restoring state enabled by PSTATE.ZA.

Removing the option, drastically simplifies the logic. Note that
I have removed the two booleans (internal to the save/restore
routines) that distinguish between having saved 'ZA' and 'ZT0',
and replaced that with a single 'VALID' bit, because I think we
can assume that the save/restore routines are called by PEs that
have the same SME state.