[sysvabi64] No BTI landing pads for absolute symbols.
It is possible to provide a static linker with branch targets
using absolute symbols. For example using --defsym. The static
cannot disassemble the location of these symbols so it isn't
possible to check whether A BTI landing pad is necessary, nor
is it always possible to place a BTI landing pad within range
of the absolute symbol address.

A user providing the static linker with function addresses
using absolute symbols is responsible for putting BTI
compatible landing pads at the start of these addresses.