Add streaming-compatible SVE variant to VFABI mangling
From the point of view of vector libraries, it is convenient to treat
SVE and streaming-compatible SVE as separate vector variants. This is
because existing optimised SVE routines may not be compatible with
streaming mode, for instance where they use SVE instructions which are
illegal in streaming mode.

This patch adds the ISA marker 'c', for streaming-compatible
SVE. Existing mapping from scalar to SVE symbols should all still make
sense with streaming-compatibility enabled, with the exception that if
the region being vectorised may have streaming enabled then the 'c'
variant should be used rather than 's'.

At present, for library purposes we are only interested in reaching a
consensus about what to name the routines, rather than extending
OpenMP and the VFABI to actually facilitate autovectorisation, however
please let me know if there is anything that I have left ambiguous or
need to add.